jour13

+++ Java/src/main/java/annees/annee2015/jour20/Annee2015Jour20Exercice1.java
@@ -18,9 +18,9 @@ public class Annee2015Jour20Exercice1 implements Runnable {
 		int total = 0;
 		while (total < input) {
 			numeroMaison++;
-			List<Integer> diviseurs = MathUtils.listeDiviseurs(numeroMaison);
+			List<Long> diviseurs = MathUtils.listeDiviseurs(numeroMaison);
 			total = 0;
-			for (Integer integer : diviseurs) {
+			for (Long integer : diviseurs) {
 				total += integer * 10;
 			}
 		}
